♻️ Moves main-window and menu up to src
