+++ src/views/example/dog-tab/index.ts
