+++ src/views/example/fun-dog/fun-dog.story.tsx
