+++ src/views/example/fun-dog/fun-dog.tsx
