+++ src/views/example/fun-dog/index.ts
