+++ src/views/example/logo/index.ts
