+++ src/views/example/long-tab/index.ts
