+++ src/views/example/welcome-screen/index.ts
