+++ src/views/example/welcome-screen/sample-tabs.tsx
