+++ src/views/example/welcome-screen/welcome-screen.tsx
